requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
gtts
SpeechRecognition
pydantic
pydantic-settings
starlette
